Fixing and organizing code
Organize everything, router, services and so on
